test(cli): extract exact dismiss notification play tests
Move the exact reviewed notification dismissal coverage from the play monolith into `game/play/notification/dismiss.test.ts` and wire the focused suite into `test:cli:play`. Keep dismissal verification fixtures local to the new owner and remove the now-unused monolith dismissal fixture branch.

Update the support modularization OpenSpec workstream to mark the exact dismiss slice complete and keep the next HUD slice blocked on its own single-writer and fixture-owner assignment.

Validation:
- bun run --cwd packages/cli test -- game/play/notification/dismiss.test.ts
- bun run --cwd packages/cli test -- game.play.test.ts game/play/notification/dismiss.test.ts game/play/notification/dismiss-queue.test.ts -t "dismiss|bulk|priorit|notifications|materializes|classifies"
- bun run check:cli
- bun run test:cli:play
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
