+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -21,8 +21,11 @@
 - [x] 2.4 Reconcile peer findings into corpus ledgers and task sequence.
 - [x] 2.5 Define shared fixture strategy before further notification/priorities
   extraction.
-- [ ] 2.6 Assign one owner at a time for `package.json` play-script wiring and
+- [x] 2.6 Assign one owner at a time for `package.json` play-script wiring and
   `packages/cli/test/commands/game.play.test.ts`.
+  - Exact dismiss-notification slice owner: support workstream owner/DRA.
+    Future slices must record their own single writer before touching either
+    file.
 - [ ] 2.7 Apply the agent framing protocol to any new or reused delegation:
   framing-design context, required skills, objective, reasoning level,
   write-set policy, `/goal` prefix for long-running work, and
@@ -51,7 +54,7 @@ runtime/direct-control claims.
 - [x] 3.9 Extract ready-unit play tests.
 - [x] 3.10 Extract notification-queue play tests.
 - [x] 3.11 Extract dismiss-notification-queue play tests.
-- [ ] 3.12 Extract exact dismiss-notification play tests.
+- [x] 3.12 Extract exact dismiss-notification play tests.
 - [ ] 3.13 Extract notification HUD materialization play tests.
 - [ ] 3.14 Extract priorities play tests.
 - [ ] 3.15 Remove residual monolith fixture ownership after the last consumer
