+++ openspec/changes/civ7-support-direct-control-modularization/workstream/cli-play-corpus.md
@@ -18,7 +18,7 @@ fixture owner.
 | Ready unit | extracted | `game/play/ready-unit.test.ts` | local/existing | test-only | Baseline evidence only. | Completed before this change. |
 | Notification queue | extracted | `game/play/notification/queue.test.ts` | local fake | test-only | Baseline evidence only. | Completed before this change. |
 | Dismiss notification queue | extracted | `game/play/notification/dismiss-queue.test.ts` | local fake | test-only | Baseline evidence only. | Completed before this change. |
-| Exact dismiss notification | `GamePlayDismissNotification` in `game.play.test.ts`: `dismisses reviewed notifications only with explicit approval reason`; `does not verify dismissal from stale nonblocking front evidence`; `does not verify dismissal from train absence while engine queue still fronts the target`; `does not verify dismissal from dismissed flag while engine queue still fronts the target` | `game/play/notification/dismiss.test.ts` | start local with exact `notificationDismissal` fake; promote only if reused by queue/direct-control atom tests | test-only, mutation-facing fake; no runtime claim | Focused filter: `dismisses reviewed notifications|does not verify dismissal`; adjacent with `game/play/notification/dismiss-queue.test.ts`; ownership scan for `GamePlayDismissNotification`, `readNotificationDismissal`, and test names. | Planned next CLI slice after fixture strategy. |
+| Exact dismiss notification | Moved from `GamePlayDismissNotification` in `game.play.test.ts`: `dismisses reviewed notifications only with explicit approval reason`; `does not verify dismissal from stale nonblocking front evidence`; `does not verify dismissal from train absence while engine queue still fronts the target`; `does not verify dismissal from dismissed flag while engine queue still fronts the target` | `game/play/notification/dismiss.test.ts` | local exact `notificationDismissal` fake | test-only, mutation-facing fake; no runtime claim | Passed: focused `bun run --cwd packages/cli test -- game/play/notification/dismiss.test.ts`; adjacent `bun run --cwd packages/cli test -- game.play.test.ts game/play/notification/dismiss.test.ts game/play/notification/dismiss-queue.test.ts -t "dismiss|bulk|priorit|notifications|materializes|classifies"`; `bun run check:cli`; `bun run test:cli:play`; ownership scan shows exact dismiss owner only in `game/play/notification/dismiss.test.ts`. | Completed in exact dismiss slice. |
 | Notification HUD materialization | `GamePlayNotifications` in `game.play.test.ts`: materializes diplomacy response, technology, culture, celebration, government, narrative, stale command-units; reads materialized notifications without sending; classifies invalid-target notices and valid reports | `game/play/notification/hud.test.ts` | named notification HUD fixture split by surface/mode; no broad 24-mode catalog | test-only | Focused filter: `materializes|classifies .* notices|classifies valid diplomatic`; adjacent with `notification/queue.test.ts`; scan `GamePlayNotifications`, `readPlayNotifications`, `playNotificationMode`. | Planned after exact dismiss. |
 | Priorities | `GamePlayPriorities` in `game.play.test.ts`: 21 compact priority/routing tests across clean-read, unit-command, exact operations, chooser options, dismissal diagnostics, ready-city blockers, and stale command-units | `game/play/priorities.test.ts` or deeper split after fixture strategy | named priority scenario builder only after HUD/ready/dismiss fixtures have owners | test-only; high relationship-label risk | Focused filter: `priorit|surfaces|routes|classifies stale`; adjacent with ready-city, ready-unit, tactical-read; scan relationship-label terms and duplicate coverage. | Planned last; do not extract before HUD. |
